auth0/requirements.txt
asgiref==3.8.1
Authlib==1.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
drf-jwt==1.19.2
idna==3.6
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
